requirements.txt
ccxt>=4.3.0
flask>=3.0.0
apscheduler>=3.10.4
pandas>=2.0.0
ta>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.1
numpy>=1.24.0
gunicorn>=21.0.0
pytrends>=4.9.2
